// frontend/src/types/api.ts
export interface User {
  id: number;
  salutation?: string;
  first_name: string;
  middle_initial?: string;
  last_name: string;
  username: string;
  email_address: string;
  contact_number?: string;
  organization_id?: number | null;
  role_id?: number | null;
  agent_id?: number | null;
  created_at: string;
  updated_at: string;
  active?: boolean;
  organization?: {
    id: number;
    organization_name: string;
    address?: string | null;
    contact_number?: string | null;
    email_address?: string | null;
  };
  role?: {
    id: number;
    role_name: string;
    description?: string;
    permissions?: string;
  };
  agent_balance?: {
    id: number;
    agent_id: number;
    balance: number;
    commission?: number;
    incentives?: number;
    quota?: number;
    incentives_value?: number;
    remarks?: string;
    created_at?: string;
    updated_at?: string;
  } | null;
}

export interface Organization {
  id: number;
  organization_name: string;
  address?: string | null;
  contact_number?: string | null;
  email_address?: string | null;
  created_at: string;
  created_by_user_id?: number | null;
  updated_at: string;
  updated_by_user_id?: number | null;
  organization_id?: number | null;
  users?: User[];
}

export interface Role {
  id: number;
  role_name: string;
  description?: string;
  permissions?: string;
  created_at: string;
  updated_at: string;
  organization_id?: number | null;
  users?: User[];
}
// ...
export interface CreateUserRequest {
  salutation?: string;
  first_name: string;
  middle_initial?: string;
  last_name: string;
  username: string;
  email_address: string;
  contact_number?: string;
  password: string;
  organization_id?: number;
  role_id?: number;
  agent_id?: number;
  darkmode?: string;
  created_by_user_id?: number | null;
  updated_by_user_id?: number | null;
  active?: number | boolean;
  commission?: number;
  quota?: number;
  incentives_value?: number;
  remarks?: string;
}

export interface UpdateUserRequest {
  salutation?: string;
  first_name?: string;
  middle_initial?: string;
  last_name?: string;
  username?: string;
  email_address?: string;
  contact_number?: string;
  password?: string;
  organization_id?: number | null | undefined;
  role_id?: number | null | undefined;
  agent_id?: number | null | undefined;
  active?: boolean;
  commission?: number;
  quota?: number;
  incentives_value?: number;
  remarks?: string;
}
// ...
export interface Agent {
    id: number;
    team_name: string;
    created_by: string;
    created_at: string;
    organization_id?: number | null;
}